refactor: Remove crawling package
